[VM] Small optimizations
- Fix high cpu usage from Headline Renderer (should really move update
to draw step)
- VMSleep gets hit a lot so tighten up its core functionality. The only
way this can really get faster now is a JIT.

(next up is SetToNext)

+++ TSOClient/tso.simantics/Marshals/VMEntityMarshal.cs
@@ -34,7 +34,7 @@ namespace FSO.SimAntics.Marshals
         public short[] Attributes;
         public VMEntityRelationshipMarshal[] MeToObject;
 
-        public uint DynamicSpriteFlags;
+        public ulong DynamicSpriteFlags;
         public LotTilePos Position;
 
         public virtual void Deserialize(BinaryReader reader)
@@ -118,7 +118,7 @@ namespace FSO.SimAntics.Marshals
             writer.Write(MeToObject.Length);
             foreach (var item in MeToObject) item.SerializeInto(writer);
 
-            writer.Write(DynamicSpriteFlags); /** Used to show/hide dynamic sprites **/
+            writer.Write((uint)DynamicSpriteFlags); /** Used to show/hide dynamic sprites **/
             Position.SerializeInto(writer);
         }
     }
